Simplify and Clarify Interfaces
- Merge Parser into Interpreter
- Reorganize project folders to make type functions clear
- Update README
