+++ Core/Helpers/StringHelper.cs
@@ -1,4 +1,4 @@
-﻿namespace Core
+﻿namespace Core.Helpers
 {
     public static class StringHelper
     {
@@ -11,13 +11,13 @@
 
                 for (int i = 0; i < str.Length; i += 2)
                 {
-                    hash1 = ((hash1 << 5) + hash1) ^ str[i];
+                    hash1 = (hash1 << 5) + hash1 ^ str[i];
                     if (i == str.Length - 1)
                         break;
-                    hash2 = ((hash2 << 5) + hash2) ^ str[i + 1];
+                    hash2 = (hash2 << 5) + hash2 ^ str[i + 1];
                 }
 
-                return hash1 + (hash2 * 1566083941);
+                return hash1 + hash2 * 1566083941;
             }
         }
         public static string[] SplitArgumentsLikeCsv(this string line, char separator = ',', bool ignoreEmptyEntries = false)
@@ -31,7 +31,7 @@
 
             if (ignoreEmptyEntries)
                 return arguments.Where(a => !string.IsNullOrWhiteSpace(a)).ToArray();
-            else 
+            else
                 return arguments;
         }
     }
