+++ Core/README.md
@@ -1,3 +1,4 @@
 ﻿# Core
 
 Contains main components for all Pure-derived executables.
+For frontend uses, just use `Interpreter`.
