+++ Frontends/Pure/Program.cs
@@ -1,4 +1,5 @@
 ﻿using Core;
+using Core.Helpers;
 
 namespace Pure
 {
@@ -7,13 +8,17 @@ namespace Pure
         static void Main(string[] args)
         {
             if (args.Length == 0)
-                new Interpreter($"""
-                    Pure Interpreter v0.0.3 (Core: {Parser.CoreVersion})
-                    """, null, null, null, null)
-                    .Start();
+            {
+                var interpreter = new Interpreter($"""
+                    Interactive Pure Interpreter (REPL) v0.0.4 (Core Version: {Interpreter.CoreVersion})
+                    This is the frontend to Pure, the scripting language.
+                    """, null, null, null, null);
+                interpreter.Start();
+                EnterInteractiveMode(interpreter);
+            }
             else if (args.Length >= 2 && (args[0] == "-m" || args[0] == "-mi"))
                 new Interpreter(string.Empty, null, args.Skip(2).ToArray(), new string[] { args[1] }, null)
-                    .Start(true);
+                    .Start();
             else if (args.Length >= 1) 
             {
                 bool interactiveMode = args.Length >= 2 && (args[0].ToLower() == "-i" || args[0].ToLower() == "--interactive");
@@ -25,7 +30,23 @@ namespace Pure
                     Console.WriteLine($"File {file} doesn't exist.");
                     return; 
                 }
-                new Interpreter(string.Empty, file, interactiveMode ? args.Skip(2).ToArray() : args.Skip(1).ToArray(), Parser.SplitScripts(File.ReadAllText(file)), file.GetDeterministicHashCode().ToString()).Start(!interactiveMode);
+                var interpreter = new Interpreter(string.Empty, file, interactiveMode ? args.Skip(2).ToArray() : args.Skip(1).ToArray(), Interpreter.SplitScripts(File.ReadAllText(file)), file.GetDeterministicHashCode().ToString());
+                interpreter.Start();
+                if(interactiveMode)
+                    EnterInteractiveMode(interpreter);
+            }
+        }
+
+        private static void EnterInteractiveMode(Interpreter interpreter)
+        {
+            while (true)
+            {
+                Console.Write(">>> ");
+                string input = Console.ReadLine().Trim();
+
+                if (input == "exit" || input == "exit()")
+                    return;
+                interpreter.Evaluate(input);
             }
         }
     }
